chore: Move Equalizer Settings "screen" into a "card"
- Follows the same idea as the "ReplayGain Pre-amp" feature.
