--- mobile/src/resources/icons/Equalizer.tsx
@@ -1,13 +1,0 @@
-import Svg, { Path } from "react-native-svg";
-
-import { useColor } from "~/modules/customization/theme/hooks";
-import type { Icon } from "./type";
-
-export function Equalizer({ size = 24, color }: Icon) {
-  const usedColor = useColor(color, "onSurface");
-  return (
-    <Svg width={size} height={size} viewBox="0 -960 960 960" fill={usedColor}>
-      <Path d="M207.94-207.96q-19.86-19.88-19.86-48.19v-124.04q0-28.12 19.89-48 19.88-19.89 48.11-19.89t48.09 19.89q19.87 19.88 19.87 48v124.04q0 28.31-19.89 48.19t-48.11 19.88q-28.23 0-48.1-19.88Zm272.04 19.88q-28.33 0-48.09-19.88-19.77-19.88-19.77-48.2v-447.91q0-28.12 19.79-47.99 19.79-19.86 48.11-19.86 28.33 0 48.19 19.88 19.87 19.88 19.87 48.01v447.91q0 28.31-19.89 48.18-19.89 19.86-48.21 19.86Zm175.85-19.88q-19.87-19.88-19.87-48.19v-251.73q0-28.13 19.89-48.01t48.11-19.88q28.23 0 48.1 19.88 19.86 19.88 19.86 48.01v251.73q0 28.31-19.89 48.19-19.88 19.88-48.11 19.88t-48.09-19.88Z" />
-    </Svg>
-  );
-}
